update product service

+++ product/src/main/java/com/lexuancong/product/viewmodel/product/ProductPreviewVm.java
@@ -2,8 +2,10 @@ package com.lexuancong.product.viewmodel.product;
 
 import com.lexuancong.product.model.Product;
 
+import java.math.BigDecimal;
+
 public record ProductPreviewVm(
-        Long id,String name,String slug,Double price ,String avatarUrl) {
+        Long id, String name, String slug, BigDecimal price , String avatarUrl) {
 
 
 }
